pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "task manager"
authors = ["deusmg <laptevap@gmail.com>"]
readme = "README.md"
packages = [
    {include = "task_manager"}
]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
dj-database-url = "^2.1.0"
gunicorn = "^21.2.0"
django-bootstrap5 = "^23.4"
whitenoise = "^6.6.0"
django-filter = "^23.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
